refactor(core): extract write coalescing and meta memory bootstrap (#214)
Made-with: Cursor
