+++ packages/memento-core/src/bootstrap.ts
@@ -14,9 +14,9 @@ import { ForgettingPolicyService } from './domains/forgetting/services/forgettin
 import { getPerformanceMonitor } from './domains/monitoring/services/performance-monitor.js';
 import { ErrorLoggingService } from './domains/monitoring/services/error-logging-service.js';
 import { PerformanceAlertService } from './domains/monitoring/services/performance-alert-service.js';
-import { WriteCoalescingManager, type CoalescedWrite } from './shared/utils/write-coalescing.js';
-import { DatabaseUtils } from './shared/utils/database.js';
+import type { WriteCoalescingManager } from './shared/utils/write-coalescing.js';
 import { createAnchorStack } from './bootstrap/anchor-stack.js';
+import { createWriteCoalescingMetaAndScore } from './bootstrap/write-and-meta.js';
 import { createMonitoringAndSchedulers } from './bootstrap/monitoring-schedulers.js';
 import { startFailureAndReflexion } from './bootstrap/failure-reflexion.js';
 import type { AnchorManager } from './domains/anchor/services/anchor/anchor-manager.js';
@@ -25,7 +25,6 @@ import type { IBatchScheduler } from './shared/interfaces/batch-scheduler.interf
 import type { IConsolidationScoreService } from './shared/interfaces/consolidation-score.interface.js';
 import type { IDatabaseOptimizer } from './shared/interfaces/database-optimizer.interface.js';
 import type { IReflexionWorker } from './shared/interfaces/reflexion-worker.interface.js';
-import { ConsolidationScoreService } from './infrastructure/consolidation-score-service.js';
 import { getVectorSearchEngine } from './domains/search/algorithms/vector-search-engine.js';
 import { getBatchScheduler } from './infrastructure/scheduler/batch-scheduler.js';
 import { SleepConsolidationService } from './domains/consolidation/services/sleep-consolidation-service.js';
@@ -34,9 +33,8 @@ import type { RelationGraphPort } from './domains/relation/ports/relation-graph.
 import { logger } from './shared/utils/logger.js';
 import { WalCheckpointScheduler } from './infrastructure/database/wal-checkpoint-scheduler.js';
 import { DatabaseLockMonitor } from './infrastructure/database/database-lock-monitor.js';
-import { MetaMemoryService } from './domains/memory/services/meta-memory-service.js';
+import type { MetaMemoryService } from './domains/memory/services/meta-memory-service.js';
 import { IntrospectionScanCache } from './domains/memory/services/introspection-scan-cache.js';
-import type { SqlParam } from './shared/types/index.js';
 import { TelemetryRepository } from './domains/telemetry/repositories/telemetry-repository.js';
 import { TelemetryService } from './domains/telemetry/services/telemetry-service.js';
 import { RuntimeDiagnosticsLogger } from './domains/monitoring/services/runtime-diagnostics-logger.js';
@@ -97,55 +95,8 @@ export async function initializeServices(db: Database.Database): Promise<ServerS
       walCheckpointScheduler,
       databaseLockMonitor
     } = await createMonitoringAndSchedulers(db);
-    let consolidationScoreService: ConsolidationScoreService | undefined;
-    const writeCoalescingManager = new WriteCoalescingManager(
-      1000,
-      async (writes: CoalescedWrite[]) => {
-        if (!db || writes.length === 0) return;
-        const currentDb = db;
-        try {
-          await DatabaseUtils.runTransaction(currentDb, async () => {
-            for (const write of writes) {
-              const updates: string[] = [];
-              const params: SqlParam[] = [];
-              if (write.fields.recall_count !== undefined) {
-                updates.push('recall_count = ?');
-                params.push(write.fields.recall_count);
-              }
-              if (write.fields.last_accessed_at !== undefined) {
-                updates.push('last_accessed_at = ?');
-                params.push(write.fields.last_accessed_at);
-              }
-              if (mementoConfig.consolidationScoreEnabled) {
-                if (write.fields.g_value !== undefined) {
-                  updates.push('g_value = ?');
-                  params.push(write.fields.g_value);
-                }
-                if (write.fields.consolidation_score !== undefined) {
-                  updates.push('consolidation_score = ?');
-                  params.push(write.fields.consolidation_score);
-                }
-              }
-              if (updates.length > 0) {
-                params.push(write.memoryId);
-                DatabaseUtils.run(
-                  currentDb,
-                  `UPDATE memory_item SET ${updates.join(', ')} WHERE id = ?`,
-                  params
-                );
-              }
-            }
-          });
-        } catch (error) {
-          logger.error(`⚠️ Write coalescing flush 실패: ${error instanceof Error ? error.message : String(error)}`);
-        }
-      }
-    );
-    if (mementoConfig.consolidationScoreEnabled) {
-      consolidationScoreService = new ConsolidationScoreService();
-    }
-    const metaMemoryService = new MetaMemoryService(db, writeCoalescingManager);
-    logger.info('MetaMemoryService 초기화 완료');
+    const { writeCoalescingManager, consolidationScoreService, metaMemoryService } =
+      createWriteCoalescingMetaAndScore(db);
     const introspectionScanCache = new IntrospectionScanCache();
     const telemetryRepository = new TelemetryRepository(db);
     const batchScheduler = getBatchScheduler();
